requirements.txt
git+git://github.com/Nihilate/Acerola.git
python-dotenv
pyTelegramBotAPI
